requirements.txt
google-generativeai
faster-whisper
SpeechRecognition
pyaudio
python-dotenv
numpy
